chore: #417 Add PayProp theme
