+++ src/styles/globals.ts
@@ -3,8 +3,7 @@ import { css } from '@linaria/core'
 import '#src/tokens/legacy-reapit/tokens.css'
 
 import '#src/tokens/dist/reapit.css'
-// TODO: Once we have the PayProp tokens, we can import them here
-// import '#src/tokens/dist/payprop.css'
+import '#src/tokens/dist/payprop.css'
 
 // export javascript colours for the SVG background in the button
 export const intentPrimary = '#4e56ea'
